--- exts/ext_asset_template/config/extension.toml
@@ -1,33 +1,0 @@
-[package]
-
-# Semantic Versioning is used: https://semver.org/
-version = "0.1.0"
-
-# Description
-category = "isaaclab"
-readme  = "README.md"
-
-title = "Extension Asset Template"
-author = "Isaac Lab Project Developers"
-maintainer = "Isaac Lab Project Developers"
-description="Extension Asset Template for Isaac Lab"
-repository = "https://github.com/isaac-sim/IsaacLabExtensionTemplate.git"
-keywords = ["extension", "asset", "template", "isaaclab"]
-
-[dependencies]
-"omni.isaac.lab" = {}
-"omni.isaac.lab_asset" = {}
-"omni.isaac.lab_tasks" = {}
-# NOTE: Add additional dependencies here
-
-[[python.module]]
-name = "ext_asset_template"
-
-[isaaclab_settings]
-# TODO: Uncomment and list any apt dependencies here.
-#       If none, leave it commented out.
-# apt_deps = ["example_package"]
-# TODO: Uncomment and provide path to a ros_ws
-#       with rosdeps to be installed. If none,
-#       leave it commented out.
-# ros_ws = "path/from/extension_root/to/ros_ws"
