--- exts/ext_asset_template/docs/CHANGELOG.rst
@@ -1,10 +1,0 @@
-Changelog
----------
-
-0.1.0 (2025-01-23)
-~~~~~~~~~~~~~~~~~~
-
-Added
-^^^^^
-
-* Created an initial assets template for building an extension or project based on Isaac Lab
