--- exts/ext_asset_template/pyproject.toml
@@ -1,3 +1,0 @@
-[build-system]
-requires = ["setuptools", "wheel", "toml"]
-build-backend = "setuptools.build_meta"
