--- exts/ext_asset_template/setup.py
@@ -1,40 +1,0 @@
-"""Installation script for the 'ext_template' python package."""
-
-import os
-import toml
-
-from setuptools import setup
-
-# Obtain the extension data from the extension.toml file
-EXTENSION_PATH = os.path.dirname(os.path.realpath(__file__))
-# Read the extension.toml file
-EXTENSION_TOML_DATA = toml.load(os.path.join(EXTENSION_PATH, "config", "extension.toml"))
-
-# Minimum dependencies required prior to installation
-INSTALL_REQUIRES = [
-    # NOTE: Add dependencies
-    "psutil",
-]
-
-# Installation operation
-setup(
-    name="ext_asset_template",
-    packages=["ext_asset_template"],
-    author=EXTENSION_TOML_DATA["package"]["author"],
-    maintainer=EXTENSION_TOML_DATA["package"]["maintainer"],
-    url=EXTENSION_TOML_DATA["package"]["repository"],
-    version=EXTENSION_TOML_DATA["package"]["version"],
-    description=EXTENSION_TOML_DATA["package"]["description"],
-    keywords=EXTENSION_TOML_DATA["package"]["keywords"],
-    install_requires=INSTALL_REQUIRES,
-    license="MIT",
-    include_package_data=True,
-    python_requires=">=3.10",
-    classifiers=[
-        "Natural Language :: English",
-        "Programming Language :: Python :: 3.10",
-        "Isaac Sim :: 2023.1.1",
-        "Isaac Sim :: 4.0.0",
-    ],
-    zip_safe=False,
-)
