+++ exts/ext_template/ext_template/assets/__init__.py
@@ -1,0 +1,6 @@
+"""
+Python module serving as a project/extension template.
+"""
+
+# Register Gym environments.
+from .assets import *
