+++ exts/ext_template/ext_template/assets/assets/__init__.py
@@ -15,13 +15,6 @@ ISAACLAB_ASSETS_EXT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__)
 ISAACLAB_ASSETS_DATA_DIR = os.path.join(ISAACLAB_ASSETS_EXT_DIR, "data")
 """Path to the extension data directory."""
 
-ISAACLAB_ASSETS_METADATA = toml.load(os.path.join(ISAACLAB_ASSETS_EXT_DIR, "config", "extension.toml"))
-"""Extension metadata dictionary parsed from the extension.toml file."""
-
-# Configure the module-level variables
-__version__ = ISAACLAB_ASSETS_METADATA["package"]["version"]
-
-
 ##
 # Configuration for different assets.
 ##
