+++ exts/ext_template/ext_template/assets/data/humanoid.usd
