+++ exts/ext_template/ext_template/assets/data/humanoid_instanceable.usd
