+++ exts/ext_template/ext_template/assets/test/test_valid_configs.py
