+++ exts/ext_template/ext_template/tasks/direct/humanoid/__init__.py
@@ -20,7 +20,7 @@ gym.register(
     entry_point=f"{__name__}.humanoid_env:HumanoidEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{__name__}.humanoid_env:HumanoidEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.humanoid_env_cfg:HumanoidEnvCfg",
         "rl_games_cfg_entry_point": f"{agents.__name__}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{agents.__name__}.rsl_rl_ppo_cfg:HumanoidPPORunnerCfg",
         "skrl_cfg_entry_point": f"{agents.__name__}:skrl_ppo_cfg.yaml",
